Add essential files in git
The Pods folder in the example is purposely not added to the git as this is a Pod project and is meant to be light weight. So cocoa pods files which are considered to be build artifacts are not added to git. But the workspace file and the pod lock file is added to git so that its easier to build the example app for testing.